Rebasing the repo build on this
To properly test multiplayer you must load the scene using the
Quickstart scene, I will fix this in future. My main Pawn system should
be well documented in the PLayerController class and the Pawn class.

if you need to find documentation for networking then just search for
"forge networking unity", this has been chosen as it best suits the
project